API documentation

Hermes functions

These functions publish and subscribe to Hermes MQTT-topics.

HermesMQTT.subscribe_to_intents — Function

subscribe_to_intents(intents, callback; moreTopics = nothing)

Subscribe to one or a list of intents and listen forever and run the callback if a matching intent is recieved.

Arguments:

intents: Abstract String or List of Abstract Strings to define intents to subscribe. The intents will be expanded to topics (i.e. "hermes/intent/SwitchOnLight")

callback: Function to be executed for a incoming message

moreTopics: keyword arg to provide additional topics to subscribe (complete names of of topics).

Details:

The callback function has the signature f(topic, intentMessage), where topic is a String and intentMessage a Dict{Symbol, Any} with the content of the payload (assuming, that the payload is in JSON-format) or a String, if the payload is not valid JSON. The callback function is spawned and the function is listening to the MQTT server while the callback is executed.

HermesMQTT.subscribe_to_topics — Function

subscribe_to_topics(topics, callback)

Subscribe to one or a list of topics and listen forever and run the callback if a matching intent is recieved.

Arguments:

topics: Abstract String or List of Abstract Strings to define topics to subscribe.

callback: Function to be executed for a incoming message

See subscribe2Intents() for details.

HermesMQTT.publish_start_session_action — Function

publish_start_session_action(text; siteID=get_siteID(),
     intentFilter=nothing, sendIntentNotRecognized=false,
     customData=nothing)

MQTT publish start session with init action

Arguments:

siteID: ID of the site in which the session is started

text: text to be said via TTS

intentFilter: Optional Array of String - a list of intent names to restrict the NLU resolution of the answer of this query.

sendIntentNotRecognized: Optional Boolean - Indicates whether the dialogue manager should handle non recognized intents by itself or sent them as an Intent Not Recognized for the client to handle.

customData: data to be sent to the service.

HermesMQTT.publish_start_session_notification — Function

publish_start_session_notification(text; siteID=get_siteID(),
                                customData=nothing)

MQTT publish start session with init notification

Arguments:

siteID: siteID

text: text to be said via TTS

customData: data to be sent to the service.

HermesMQTT.publish_end_session — Function

publish_end_session(text...; sessionID=get_sessionID())

MQTT publish end session.

Arguments:

sessionId: ID of the session to be terminated as String. If omitted, sessionId of the current will be inserted.

text: text to be said via TTS

HermesMQTT.publish_hotword_on — Function

function publish_hotword_on(siteID)

Publish a hotword-on topic for the siteID.

Arguments:

siteID: mandatory argument siteID must be given as String (no default)

HermesMQTT.publish_hotword_off — Function

function publish_hotword_off(siteID)

Publish a hotword-off topic for the siteId.

Arguments:

siteID: mandatory argument siteId must be given as String (no default)

Dialogue manager functions

In addition to functions to work with the dialogue manager, advanced direct dialogues are provided that can be included in the control flow of the program.

HermesMQTT.publish_continue_session — Function

publish_continue_session(text...; sessionID=get_sessionID(),
     intentFilter = nothing,
     customData = nothing, sendIntentNotRecognized = false)

MQTT publish continue session.

Arguments:

sessionID: ID of the current session as String

text: text to be said via TTS

intentFilter: Optional Array of String - a list of intents names to restrict the NLU resolution on the answer of this query.

customData: Optional String - an update to the session's custom data.

sendIntentNotRecognized: Optional Boolean - Indicates whether the dialogue manager should handle non recognized intents by itself or sent them as an Intent Not Recognized for the client to handle.

Missing docstring.

Missing docstring for listen_intents_one_time. Check Documenter's build log for details.

HermesMQTT.configure_intent — Function

configure_intent(intent, on)

Enable or disable an intent.

intent: one intent to be configured

on: boolean value; if true, the intent is enabled; if false it is disabled.

Missing docstring.

Missing docstring for ask_yes_or_no_or_unknown. Check Documenter's build log for details.

HermesMQTT.ask_yes_or_no — Function

ask_yes_or_no(question)

Ask the question and listen to the intent "ADoSnipsYesNoDE" and return :true if "Yes" or "No" otherwise.

Arguments:

question: String with the question to uttered

HermesMQTT.publish_say — Function

publish_say(text...; sessionID=get_sessionID(), siteID=nothing,
                lang=nothing, id=nothing, wait=true)

Let the TTS say something.

The variant with a Symbol as first argument looks up the phrase in the dictionary of phrases for the selected language by calling getText().

Arguments:

text: text to be said via TTS

lang: optional language code (i.e. en-US) to use when saying the text. If not specified, default language will be used

sessionId: optional ID of the session if there is one

id: optional request identifier. If provided, it will be passed back in the response on hermes/tts/sayFinished.

wait: wait until the massege is spoken (i.e. wait for the MQTT-topic)

HermesMQTT.publish_nlu_query — Function

publishnluquery(input; siteID="default", sessionID=nothing, intentFilter=nothing)

Publish a NLU request for the command input.

HermesMQTT.publish_intent — Function

publish_intent(payload, topic=nothing)

Publish an intent with the payload as topic.

Interaction with assistant Rhasspy/Susi/Snips

HermesMQTT.get_language — Function

get_language()

Return the language in the Module HermesMQTT

HermesMQTT.upload_intents — Function

upload_intents(skill_name=nothing)

Upload the intents and slots for the given skill to Rhasspy.

Arguments:

skill_name: name of the skill. The name must match the name of the skill directory to make it possible to find the intents and slots in the directory <skillname>/profiles/<lang>/.

Functions to handle intents

HermesMQTT.is_false_detection — Function

is_false_detection(payload)

Return true, if the current intent is not correct for the uttered command (i.e. false positive).

Arguments:

payload: Dictionary with the payload of a recognised intent.

Details:

All lines of the config.ini are analysed, witch match expressions:

`<intentname>:must_include:<description>=<list of words>

An example would be:

switchOnOff:must_include = on, light

switchOnOff:must_chain = light,on

switchOnOff:must_span = switch,light,on

switchOnOff:must_span = switch,on

The command must include all words in the correct order of at least one parameter lines and the words must span the complete line (i.e. the command starts with first word and ends with the last word of the list).

Several lines are possible; the last part of the parameter name is used as description and to make the parameter names unique.

config.ini functions

Helper functions for read values from the file config.ini.

config.ini files follow the normal rules as for all Snips apps, with some extensions:

spaces are allowed around the =

the parameter value may not contain whitespace; i.e. light=my light is better written as light=my_light

if the value of the parameter (right side) includes commas, the value is interpreted as a comma-separated list of values. In this case, the reader-function will return an array of Strings with the values (which can be accessed by their index).

parameter names may have a prefix (set by the function set_config_prefix()). If set, all config-functions will try to find parameter names with prefix. Example: If the config.ini includes the lines:

main_light:ip=192.168.0.15
wall_light:ip=192.168.0.16

the following code returns 192.168.0.15 in the first call and 192.168.0.16 in the second. This makes it easy to delegate config-reads to sub-functions:

set_config_prefix("main_light")
main_ip = get_config("ip")

set_config_prefix("wall_light")
wall_ip = get_config("ip")

Obviously is possible to access the parameters directly via get_config("main_light:ip") or get_config(Symbol("main_light:ip")) without setting the prefix (the second form will work even if another prefix is set; see doc of the functions for details).

HermesMQTT.get_config_skill — Function

get_config(name; multiple = false, one_prefix = nothing)
get_config_skill(name; multiple = false, one_prefix = nothing,
                skill=get_appname())

Return the parameter value of the config.ini with name or nothing if the param does not exist. Return value is of type AbstractString, if it is a single value or of type AbstractArray{AbstractString} if the a list of values is read.

Arguments:

name: name of the config parameter as Symbol or String

multiple: if true an array of values is returned, even if only a single value have been read.

one_prefix: if defined, the prefix will be used only for this single call instead of the stored prefix.

Details:

If name is of type Symbol, it is treated as key in the Dirctionary of parameter values. If name is an AbstractString, the prefix is added if a prefix is defined (as <prefix>:<name>). 'get_config()' returns ''nothing if something is wrong.

config.ini entries are stored with the corresponding skill name (appname), The function get_config() calls get_config_skill()with the current appname used as skill. If a config-entry for a different skill is wanted, the functiongetconfigskill(...; skill="skillname")` can be used.

HermesMQTT.is_in_config_skill — Function

is_in_config(name)
is_in_config_skill(name; skill=get_appname())

Return true if a parameter with name exists.

Arguments:

name: name of the config parameter as Symbol or String

HermesMQTT.match_config_skill — Function

match_config(name::Symbol, val::String)
match_config_skill(name::Symbol, val::String; skill=get_appname())

Return true if the parameter with name name of the config.ini has the value val or one element of the list as the value val.

Arguments:

name: name of the config parameter as Symbol or String

val: desired value

HermesMQTT.set_config_prefix — Function

set_config_prefix(new_prefix)

Set the prefix for all following calls to a parameter from config.ini. All parameter names, gives as Strings will be modified as <prefix>:<name>.

Details:

Do not forget to reset the prefix (reset_config_prefix()) as possible, to avoid side effects as the prefix is applied globally!

HermesMQTT.reset_config_prefix — Function

reset_config_prefix()

Remove the prefix for all following calls to a parameter from config.ini.

HermesMQTT.get_all_config — Function

get_all_config()

Return a Dict with the complete config.ini.

HermesMQTT.read_config — Function

read_config(app_dir; skill=get_appname())

Read the lines of the App's config file and return a Dict with config values.

Arguments:

appDir: Directory of the currently running app.

HermesMQTT.is_config_valid — Function

is_config_valid(name; regex = r".", elem = 1, errorMsg = ERRORS_EN[:error_config],, 
                skill=get_appname())

Return true, if the parameter name have been read correctly from the config.ini file and false otherwise. By default "correct" means: it is a String with length > 0. For a moore specific test, a regex can be provided.

Arguments:

name: name of parameter as AbstractString or Symbol

regex: optional regex for the test

elem: element to be tested, if the parameter returns an array

errorMsg: alternative error message.

HermesMQTT.get_config_path — Function

get_config_path(name, default_path; one_prefix = nothing,
                skill=get_appname())

Read the config value 'name' as filename and generate a full (absolute) path:

if fName starts with '/', it is returned as is.

otherwise a full path is created with 'defaultPath' as prefix.

Arguments:

name: name of the config parameter as Symbol or String

default_path: path to be used if name is not already a path

one_prefix: if defined, the prefix will be used only for this single call instead of the stored prefix.

HermesMQTT.load_hermes_config — Function

load_hermes_config(hermes_dir)

Load config setting for the HermesMQTT framework.

Arguments:

hermes_dir: path to the HermesMQTT-config.ini .../Skills/HermesMQTT.jl

HermesMQTT.load_skill_config — Function

load_skill_config(app_dir; skill=get_appname()

Load config setting for a skill.

Arguments:

app_dir: path to the config.ini .../Skills/Skill (/config.ini)

HermesMQTT.load_two_configs — Function

load_two_configs(app_dir, hermes_dir=nothing, skill=get_appname())

Load config setting for a skill in dir app_dir and config of the HermesMQTTframework fromapp_dir/../HermesMQTT/config.ini` or the specified dir.

Arguments:

hermes_dir: path to the HermesMQTT-config.ini .../Skills/HermesMQTT.jl

Slot access functions

Functions to read values from slots of recognised intents.

HermesMQTT.extract_slot_value — Function

extract_slot_value(slot_name, payload; default=nothing, as=String)

Return the value of a slot.

Nothing is returned, if

no slots in payload,

no slots with name slot_name in payload,

no value in slot slot_name,

or the value cannot be casted into the type, specified ba the as= argument.

Arguments:

slot_name: name of the slot as defined in the intent

payload: JSON payload

default: default value; returned if the slot is not present in the payload

as: type of the returned value (default: String)

HermesMQTT.extract_multislot_values — Function

extract_multislot_values(slot_names,payload)

Return a list with all values of a list of slots.

Nothing is returned, if

no slots in payload,

no slots with name in slotNames in payload,

no values in slot slotNames.

HermesMQTT.is_in_slot — Function

is_in_slot(slot_name, value, payload)

Return true, if the value is present in the slot slotName of the JSON payload (i.e. one of the slot values must match). Return false if something is wrong (value not in payload or no slots slotName.)

HermesMQTT.is_on_off_matched — Function

is_on_off_matched(payload, device_name; siteID=get_siteID())

Action to be combined with the Hermes ON/OFF-intent Susi:on_off. Depending on the payload the function returns:

:on if "on"

:off if "off"

:matched, if the device is matched but no on or off

:unmatched, if one of

wrong siteId

wrong device

Arguments:

payload: payload of intent

siteId: siteId of the device to be matched with the payload of intent if siteId == "any", the device will be matched w/o caring about siteId or room.

device_name : name of device to be matched with the payload of intent

HermesMQTT.is_valid_or_end — Function

is_valid_or_end(param; error_msg = "mandatory parameter is nothing")

End the session, with the message, if the param is nothing and returns false or true otherwise.

Function is a shortcut:

if isnothing(param)
    publishEndSession(:error)
    return true
end

is:

is_valid_or_end(param, :error) || return true

Arguments:

param: any value (from slot or config.ini) that may be nothing

errorMsg: Error as string or key of a text in the languages Dict (::Symbol).

HermesMQTT.read_time_from_slot — Function

read_time_from_slot(slot_name, payload)

Return a DateTime from a slot with a time string of the format "2019-09-03 18:00:00 +00:00" or nothing if it is not possible to parse the slot.

MQTT functions

Low-level API to MQTT messages (publish and subscribe). In the QuickAndDirty framework, these functions are calling Eclipse mosquitto_pub and mosquitto_sub. However this first (and preliminary) implementation is surpriningly robust and easy to maintain - there might be no need to change.

HermesMQTT.subscribe_MQTT — Function

subscribe_MQTT(topics, callback)

Listen to one or more topics.

Arguments

topics: AbstractString or List of AbtsractString to define topics to subscribe

callback: Function to be executed for a incoming message.

Details:

The callback function has the signature f(topic, payload), where topic is a String and payload a Dict{Symbol, Any} with the content of the payload (assuming, that the payload is in JSON-format) or a String, if the payload is not a valid JSON.

The callback function is spawned.

HermesMQTT.read_one_MQTT — Function

read_one_MQTT(topics; timeout=TIMEOUT)

Listen to one or more topics until one message is retrieved and return topic as string and payload as Dict or as String if JSON parsing is not possible).

Arguments

topics: AbstractString or List of AbstractString to define topics to subscribe

HermesMQTT.publish_MQTT — Function

publish_MQTT(topic, payload; file=false)

Publish a MQTT message.

Arguments

topics: String with the topic

payload: Dict() with message

file: if true, the file will be sent (via -f payload, otherwise the payload will be sent as JSON string via -m)

HermesMQTT.publish_MQTT_file — Function

publish_MQTT_file(topic, fname)

Publish a MQTT message with a file as payload.

Arguments

topics: String with the topic

fname: full path and name of file to be published

Handle background information of recognised intent

HermesMQTT.set_siteID — Function

set_siteID(siteID)

Set the siteID in the Module HermesMQTT (necessary to direct the say() output to the current room)

HermesMQTT.get_siteID — Function

get_siteID()

Return the siteID in the Module HermesMQTT (necessary to direct the say() output to the current room)

HermesMQTT.set_sessionID — Function

set_sessionID(sessionID)

Set the sessionId in the Module SnipsHermesQnD. The sessionId will be used to publish Hermes messages inside a runing session. The framework handles this in the background.

Arguments:

sessionId: as String from a Hermes payload.

HermesMQTT.get_sessionID — Function

get_sessionID()

Return the sessionId of the currently running session.

HermesMQTT.set_module — Function

set_module(currentModule)

Set the module of the currently running app in SnipsHermesQnD. The framework uses this in the background.

Arguments:

currentModule: The module in which the current skill is running. (acessible via marco @__MODULE__)

HermesMQTT.get_module — Function

get_module()

Return the module of the currently running app.

HermesMQTT.set_appdir — Function

set_appdir(app_dir)

Store the directory appdir as CURRENT[:app_dir] in the current session

HermesMQTT.get_appdir — Function

get_appdir()

Return the directory of the currently running app

HermesMQTT.set_appname — Function

set_appname(app_name)

Store the name of the current app/module in CURRENT.

HermesMQTT.get_appname — Function

get_appname()

Return the name of the currently running app.

HermesMQTT.set_topic — Function

set_topic(topic)

Set the topic for which the currently running app is working. The framework uses this in the background.

Arguments:

topic: name of current topic

HermesMQTT.get_topic — Function

get_topic()

Return the topic of the currently running app.

HermesMQTT.set_intent — Function

set_intent(payload)

Set the intent for which the currently running app is working. The framework uses this in the background.

Arguments:

topic: name of current topic

HermesMQTT.get_intent — Function

get_intent()
get_intent(payload)

Return the intent name of the currently running app.

Multi-language utilities

HermesMQTT.set_language — Function

set_language(lang)

Set the language in the Module HermesMQTT

HermesMQTT.add_text — Function

add_text(lang::AbstractString, key, text)

Add the text to the dictionary of text sniplets for the language lang and the key key.

Arguments:

lang: String with language code ("en", "de", ...)

key: Symbol or string with unique key for the text

text: String or array of String with the text(s) to be uttered.

Details:

If text is an Array, all texts will be saved and the function J4H.lang_text() will return a randomly selected text from the list.

If the key already exists, the new text will be added to the the Array of texts for a key.

HermesMQTT.lang_text — Function

function lang_text(texts..., lang=get_language())

Join the elements of texts into a single string and replace each string by the sentence of the language.

Arguments:

texts: iterable collection of elements that can be converted into strings. Symbols will be replaced by the languange snipplets.

Hardware control

Some devices can be directly controlled by the framework. In order to stay in the style of Snips, it is possible to run Shelly WiFi-switches without any cloud accounts and services. The Shelly-devices come with an own WiFi network. After installing the device just connect to Shelly's access point (somthing like shelly1-35FA58) and configure the switch for DHCP in your network with teh selft-explaining the web-interface of the device. At no point it's necessary to create an account or use a cloud service (although the Shelly1 documentation recommends).

HermesMQTT.set_GPIO — Function

set_GPIO(gpio, onoff::Symbol)

Switch a GPIO on or off with pigs.

Arguments:

gpio: ID of GPIO (not pinID)

onoff: one of :on or :off

HermesMQTT.switch_shelly_1 — Function

switch_shelly_1(ip, action; timer = 10)

Switch a shelly1 device with IP ip on or off, depending on the value given as action and return true, if successful.

Arguments:

ip: IP address or DNS name of Shelly1 device

action: demanded action as symbol; one of :on, :off, push or :timer. action :push will switch on for 200ms to simulate a push. action :timer will switch off after timer secs.

timer: if action == timer, the device is switched on and off again after timer secs (default 10s).

For the API-doc of the Shelly devices see: <https://shelly-api-docs.shelly.cloud>.

HermesMQTT.switch_shelly_25_relay — Function

switch_shelly_25_relay(ip, relay, action; timer = 10)

Switch the relay relay of a shelly2.5 device with IP ip on or off, depending on the value given as action and return true, if successful.

Arguments:

ip: IP address or DNS name of Shelly2.5 device

relay: Number of relay to be switched (0 or 1)

action: demanded action as symbol; one of :on, :off, push or :timer. action :push will switch on for 200ms to simulate a push. action :timer will switch off after timer secs.

timer: if action == timer, the device is switched on and off again after timer secs (default 10s).

For the API-doc of the Shelly devices see: <https://shelly-api-docs.shelly.cloud>.

HermesMQTT.move_shelly_25_roller — Function

move_shelly_25_roller(ip, action; pos = 100, duration = 1)

Move a roller with a shelly2.5 device with IP ip and return true, if successful.

Arguments:

ip: IP address or DNS name of Shelly2.5 device

action: demanded action as symbol; one of :open, :close, :stop or :to_pos.

pos: desired position in percent

'duration': is not yet implemented.

For the API-doc of the Shelly devices see: <https://shelly-api-docs.shelly.cloud>.

Status database

The framework handles a database to save status about house and devices, controlled by the assistant. The database is stored on disk in order to persist in case of a system crash or restart.

Every skill can store and read Dicts() as entries with a unique key or values as field-value-pairs as part of an entry.

The db looks somehow like:

{
    "irrigation" :
    {
        "time" : "2019-08-26T10:12:13.177"
        "writer" : "Susi:Irrigation",
        "payload" :
        {
            "status" : "on",
            "next_status" : "off"
        }
    }
}

Location of the database file is <application_data_dir>/HermesMQTT/<database_file> where application_data_dir and database_file> are parameters in the config.ini of the framework.

Missing docstring.

Missing docstring for db_write_payload. Check Documenter's build log for details.

HermesMQTT.db_write_value — Function

db_write_value(key, field, value)

Write a field=>value pair to the entry of a database entry. The field is overwitten if the entry already exists, or created elsewise. The database is written to the JSON-file after the write.

Arguments

key: unique key of the database entry of type AbstractString or Symbol

field: database field of the entry with key key to be written (AbstractString or Symbol).

value: value to be stored in the field, typically a Dict() will be stored as JSON.

HermesMQTT.db_has_entry — Function

db_has_entry(key)

Check if the database has an entry with the key key and return true or false otherwise.

Arguments

key: unique key of the database entry of type AbstractString or Symbol

HermesMQTT.db_read_entry — Function

db_read_entry(key)

Read the complete entry with the key key from the status database and return the entry as Dict() or nothing if not in the database.

Arguments

key: unique key of the database entry of type AbstractString or Symbol

HermesMQTT.db_read_value — Function

db_read_value(key, field)

Read the field field of the entry with the key key from the status database and return the value or nothing if not in the database.

Arguments

key: unique key of the database entry of type AbstractString or Symbol

field: database field of the payload of the entry with key key (AbstractString or Symbol).

Scheduler

The framework provides a scheduler which allows to execute intents or commands at a specified time in the future.

HermesMQTT.publish_schedule_command — Function

publish_schedule_command(command, exec_time, origin; 
                siteID=get_site_id(), sessionID=mk_session_id())

Publish a command to be executed at a specific time.

Utility functions

Little helpers to provide functionality which is commonly needed when developing a skill.

HermesMQTT.readable_date_time — Function

function readable_date_time(datetime::AbstractString; lang=get_language())
                          wholeDay = false, onlyDay = false)
function readable_date_time(datetime::DateTime; lang=get_language(),
                          wholeDay = false, onlyDay = false)
function readable_date(datetime::Date; lang=get_language())

Return human readable date and time, like "Friday, 1. January 2018", "9 15"

Supported languages: "en", "de", "fr".

Arguments:

datetime: date and optional time as ISO string or as DateTime object.

lang: Language code (e.g. "en" or "de")

wholeDay: boolean, if true, no time is returned, instead "whole day"

onlyDay: just tell the day, not the time

onlyTime: just tell the time, not the day

Value:

String value with readable date and time.

HermesMQTT.get_weather — Function

get_weather()

Return a Dict with weather information from openweather.org or weatherapi.com. The config.ini of the framework must include the lines to define which service to be used and api key and location for the service used (available from openweather.org or weatherapi.com).

# weather_service=openweather
weather_service=weatherapi
openweather:api_key=insert_valid_API-key_here
openweather:city_id=6350865

weatherapi:api_key=insert_valid_API-key_here
weatherapi:location=52.52,13.40

Value:

The return value has the elements:

:service: name of the weather service

:pressure: pressure in hPa

:temperature: temperature in degrees Celsius

:windspeed: wind speed in meter/sec

:winddir: wind direction in degrees

:clouds: cloudiness in percent

:rain: rain forecast for today

:rain1h

:rain3h: rain in mm in the last 1 or 3 hours

:sunrise

:sunset: local time of sunrise/sunset as DateTime object

If something went wrong with the API-service, nothing is returned.

HermesMQTT.tryrun — Function

tryrun(cmd; wait = true, error_msg = ERRORS_EN[:error_script], silent = false)

Try to run an external command and returns true if successful or false if not.

Arguments:

cmd: command to be executed on the shell

wait: if true, wait until the command has finished

error_msg: AbstractString or key to multi-language dict with the error message.

silent: if true, no error is published, if something went wrong.

HermesMQTT.ping — Function

ping(ip; c = 1, W = 1)

Return true, if a ping to the ip-address (or name) is successful.

Arguments:

c: number of pings to send (default: 1)

W: timeout (default 1 sec)

HermesMQTT.try_read_textfile — Function

try_read_textfile(fname, error_msg = TEXTS[:error_read])

Try to read a text file from file system and return the text as String or an String of length 0, if something went wrong.

HermesMQTT.try_parse_JSON_file — Function

try_parse_JSON_file(fname; quiet = false)

Parse a JSON file and return a hierarchy of Dicts with the content.

keys are changed to Symbol

on error, an empty Dict() is returned

Arguments:

fname: filename

quiet: if false Snips utters an error message

HermesMQTT.try_parse_JSON — Function

try_parse_JSON(text)

parses a JSON and returns a hierarchy of Dicts{Symbol, Any} and Arrays with the content or a string (text), if text is not a valid JSON, the raw string is returned.

Missing docstring.

Missing docstring for try_mk_JSON. Check Documenter's build log for details.

Missing docstring.

Missing docstring for print_log. Check Documenter's build log for details.

Missing docstring.

Missing docstring for print_debug. Check Documenter's build log for details.

Index

HermesMQTT.add_text

HermesMQTT.ask_yes_or_no

HermesMQTT.configure_intent

HermesMQTT.db_has_entry

HermesMQTT.db_read_entry

HermesMQTT.db_read_value

HermesMQTT.db_write_value

HermesMQTT.extract_multislot_values

HermesMQTT.extract_slot_value

HermesMQTT.get_all_config

HermesMQTT.get_appdir

HermesMQTT.get_appname

HermesMQTT.get_config_path

HermesMQTT.get_config_skill

HermesMQTT.get_intent

HermesMQTT.get_language

HermesMQTT.get_module

HermesMQTT.get_sessionID

HermesMQTT.get_siteID

HermesMQTT.get_topic

HermesMQTT.get_weather

HermesMQTT.is_config_valid

HermesMQTT.is_false_detection

HermesMQTT.is_in_config_skill

HermesMQTT.is_in_slot

HermesMQTT.is_on_off_matched

HermesMQTT.is_valid_or_end

HermesMQTT.lang_text

HermesMQTT.load_hermes_config

HermesMQTT.load_skill_config

HermesMQTT.load_two_configs

HermesMQTT.match_config_skill

HermesMQTT.move_shelly_25_roller

HermesMQTT.ping

HermesMQTT.publish_MQTT

HermesMQTT.publish_MQTT_file

HermesMQTT.publish_continue_session

HermesMQTT.publish_end_session

HermesMQTT.publish_hotword_off

HermesMQTT.publish_hotword_on

HermesMQTT.publish_intent

HermesMQTT.publish_nlu_query

HermesMQTT.publish_say

HermesMQTT.publish_schedule_command

HermesMQTT.publish_start_session_action

HermesMQTT.publish_start_session_notification

HermesMQTT.read_config

HermesMQTT.read_one_MQTT

HermesMQTT.read_time_from_slot

HermesMQTT.readable_date_time

HermesMQTT.reset_config_prefix

HermesMQTT.set_GPIO

HermesMQTT.set_appdir

HermesMQTT.set_appname

HermesMQTT.set_config_prefix

HermesMQTT.set_intent

HermesMQTT.set_language

HermesMQTT.set_module

HermesMQTT.set_sessionID

HermesMQTT.set_siteID

HermesMQTT.set_topic

HermesMQTT.subscribe_MQTT

HermesMQTT.subscribe_to_intents

HermesMQTT.subscribe_to_topics

HermesMQTT.switch_shelly_1

HermesMQTT.switch_shelly_25_relay

HermesMQTT.try_parse_JSON

HermesMQTT.try_parse_JSON_file

HermesMQTT.try_read_textfile

HermesMQTT.tryrun

HermesMQTT.upload_intents